Adds an overload to SendAsync which will accept an object (as long as it is not of the typed TResponse type).
It will check to see if an IResponseInterceptor has been provided to the Definition of the endpoint, and if it has, run that code.

Then it checks to see if a response has started, if it has, it will simply return, or if not, it will run the SendAsync extension of HttpContext as normal.